// src/app/models/institution-model.ts
import { Campus } from "./campus-model";

export class Institute {
    InstitutionId: number;
    InstName: string;
    InstAlias: string;
    PartnerTypeId:number;
    InstLogoPath: string;
    InstImagePath: string;
    InstAddress1: string;
    InstAddress2: string;
    InstCity: number;
    InstProvince: number;
    InstCountry: number;
    InstZipCode: string;
    InstAbout: string;
    InstType: number;
    InstFounded: string;
    InstTotalStudents: number;
    InstLocationGEOcode: string;
    InstAvgCostTuition_Year: string;
    InstCostofLiving_Year: string;
    AddUserId: number;
    InstCityName: string;
    InstCountryName: string;
    InstTypeName: string;
    InstApplicationFee:string;
    DLINo:string;
    InstFeatures:string;
    InstCurrency:string;
    PaymentLink:string;
    Website:string;
    InvoiceFormat:string;
    LetterHeadPath:string;
    LocalRank?:number;
    GlobalRank?:number;
    InstVideo?:string;
    HowToReach?:string;
    isCreditCardRequired?:boolean;
}